\name{prepotol-package}
\alias{prepotol-package}
\alias{prepotol}
\docType{package}
\title{
\packageTitle{prepotol}
}
\description{
\packageDescription{prepotol}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{prepotol}
\packageIndices{prepotol}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{prepotol}

Maintainer: \packageMaintainer{prepotol}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
